+++ app/src/main/AndroidManifest.xml
@@ -19,10 +19,6 @@
             android:name=".SignUpActivity"
             android:theme="@style/AppTheme.NoActionBar.Light" />
         <activity
-            android:name=".MainActivity"
-            android:label="@string/title_activity_main"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity
             android:name=".LoginActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar.Light" >
