--- app/src/main/java/com/example/chotuvemobileapp/MainActivity.kt
@@ -1,82 +1,0 @@
-package com.example.chotuvemobileapp
-
-import android.content.Intent
-import android.net.Uri
-import android.os.Bundle
-import android.provider.OpenableColumns
-import android.view.View
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.google.android.material.snackbar.Snackbar
-import com.google.firebase.storage.FirebaseStorage
-import com.google.firebase.storage.StorageReference
-import kotlinx.android.synthetic.main.activity_main.*
-
-
-class MainActivity : AppCompatActivity() {
-    var uri = null as Uri?
-    private lateinit var mStorageRef : StorageReference
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        setSupportActionBar(toolbar)
-        fab.setOnClickListener { view ->
-            Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                .setAction("Action", null).show()
-        }
-
-        mStorageRef = FirebaseStorage.getInstance().reference
-    }
-
-    fun onClickAddVideo(view: View) {
-        val intent = Intent(Intent.ACTION_GET_CONTENT).apply {
-            type = "video/*"
-        }
-
-        startActivityForResult(Intent.createChooser(intent, "Select video"), 0)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode == RESULT_OK) {
-            uri = data!!.data
-        }
-
-        val mUri = uri!!
-        val riversRef : StorageReference = mStorageRef.child(getFileName(mUri))
-
-        riversRef.putFile(mUri)
-            .addOnSuccessListener { taskSnapshot -> // Get a URL to the uploaded content
-                Toast.makeText(applicationContext, "OK updated", Toast.LENGTH_LONG).show()
-                // val downloadUrl: Uri = taskSnapshot.getDownloadUrl()
-            }
-            .addOnFailureListener { Exception ->
-                val errorMessage = Exception.message
-                Toast.makeText(applicationContext, errorMessage, Toast.LENGTH_LONG).show()
-            }
-    }
-
-    private fun getFileName(uri: Uri) : String {
-        var result = null as String?
-        if (uri.scheme.equals("content")) {
-            val cursor = contentResolver.query(uri, null, null, null, null)
-            try {
-                if (cursor != null && cursor.moveToFirst()) {
-                    result = cursor.getString(cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME))
-                }
-            } finally {
-                cursor!!.close()
-            }
-        }
-
-        if (result == null) {
-            result = uri.path;
-            val cut = result!!.lastIndexOf('/')
-            if (cut != -1) {
-                result = result.substring(cut + 1)
-            }
-        }
-        return result
-    }
-}
