+++ app/src/main/java/com/example/chotuvemobileapp/ui/home/HomeFragment.kt
@@ -1,19 +1,28 @@
 package com.example.chotuvemobileapp.ui.home
 
+import android.app.Activity.RESULT_OK
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
+import android.provider.OpenableColumns
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProviders
 import com.example.chotuvemobileapp.R
+import com.google.firebase.storage.FirebaseStorage
+import com.google.firebase.storage.StorageReference
+import kotlinx.android.synthetic.main.fragment_home.*
 
 class HomeFragment : Fragment() {
 
     private lateinit var homeViewModel: HomeViewModel
+    var uri = null as Uri?
+    private lateinit var mStorageRef : StorageReference
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -27,6 +36,64 @@ class HomeFragment : Fragment() {
         homeViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
+        mStorageRef = FirebaseStorage.getInstance().reference
+
+
         return root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        AddVideo.setOnClickListener{
+            val intent = Intent(Intent.ACTION_GET_CONTENT).apply {
+                type = "video/*"
+            }
+
+            startActivityForResult(Intent.createChooser(intent, "Select video"), 0)
+        }
+    }
+
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (resultCode == RESULT_OK) {
+            uri = data!!.data
+        }
+
+        val mUri = uri!!
+        val riversRef : StorageReference = mStorageRef.child(getFileName(mUri))
+
+        riversRef.putFile(mUri)
+            .addOnSuccessListener { taskSnapshot -> // Get a URL to the uploaded content
+                Toast.makeText(context, "OK updated", Toast.LENGTH_LONG).show()
+                // val downloadUrl: Uri = taskSnapshot.getDownloadUrl()
+            }
+            .addOnFailureListener { Exception ->
+                val errorMessage = Exception.message
+                Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show()
+            }
+    }
+
+    private fun getFileName(uri: Uri) : String {
+        var result = null as String?
+        if (uri.scheme.equals("content")) {
+            val cursor = requireActivity().contentResolver.query(uri, null, null, null, null)
+            try {
+                if (cursor != null && cursor.moveToFirst()) {
+                    result = cursor.getString(cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME))
+                }
+            } finally {
+                cursor!!.close()
+            }
+        }
+
+        if (result == null) {
+            result = uri.path;
+            val cut = result!!.lastIndexOf('/')
+            if (cut != -1) {
+                result = result.substring(cut + 1)
+            }
+        }
+        return result
     }
 }
